Relink local plugins on update
When running `packager#update()`, if a local plugin has already been linked, the linking operation will exit with a status code of 1. To prevent this I've added the `f` flag to the `ln` command which will remove the link if it already exists, and then make a new link. This not only prevents the error on update, but also handles the case where the target of the link has moved or changed.